refactor: broke up player_assets into separate modules
