+++ website/api/http.py
@@ -16,10 +16,13 @@ import website.utils.misc
 import website.utils.network_helpers
 import website.utils.resource_market
 from website.config.assets import wind_power_curve
+from website.database.active_facility import ActiveFacility
 from website.database.map import Hex
 from website.database.messages import Chat
+from website.database.ongoing_construction import OngoingConstruction
 from website.database.player import Network, Player
-from website.database.player_assets import ActiveFacility, OngoingConstruction, ResourceOnSale, Shipment
+from website.database.resource_on_sale import ResourceOnSale
+from website.database.shipment import Shipment
 from website.game_engine import Confirm, GameException
 from website.technology_effects import get_current_technology_values
 from website.utils.assets import package_projects_data
