+++ website/database/resource_on_sale.py
@@ -1,0 +1,12 @@
+from website import db
+
+
+class ResourceOnSale(db.Model):
+    """Class that stores resources currently on sale"""
+
+    id = db.Column(db.Integer, primary_key=True)
+    resource = db.Column(db.String(10))
+    quantity = db.Column(db.Float)
+    price = db.Column(db.Float)
+    creation_date = db.Column(db.DateTime)
+    player_id = db.Column(db.Integer, db.ForeignKey("player.id"))  # can access player directly with .player
