+++ website/utils/assets.py
@@ -7,8 +7,9 @@ from typing import List
 from flask import current_app
 
 from website import db, game_engine, technology_effects
+from website.database.active_facility import ActiveFacility
+from website.database.ongoing_construction import OngoingConstruction
 from website.database.player import Player
-from website.database.player_assets import ActiveFacility, OngoingConstruction
 from website.game_engine import Confirm, GameEngine, GameException
 from website.utils.network_helpers import reorder_facility_priorities
